requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
PyMuPDF>=1.23.8
